fix strings incorrectly being treated as enums
